+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.0.2] - 2016-05-18
+### Added
+* missing dependency snakeyaml-1.1.12
+* classicmodels database schema and data for testing
+* base command builder
+
+### Updated
+* mysql-connector-java-5.1.39-bin.jar
+
 ## [0.0.1] - 2016-05-11
 ### Added
-- skeleton for Liquibase update command
+* skeleton for Liquibase update command